Move handleUiCommands to viewmodel
